+++ Sources/BugsnagPerformance/Public/BugsnagPerformanceSpan.mm
@@ -46,6 +46,8 @@ static CFAbsoluteTime currentTimeIfUnset(CFAbsoluteTime time) {
     if ((self = [super initWithTraceId:traceId spanId:spanId])) {
         _actuallyStartedAt = CFAbsoluteTimeGetCurrent();
         _actuallyEndedAt = CFABSOLUTETIME_INVALID;
+        // Ensure endAbsTime is marked invalid until an end time is set.
+        _endAbsTime = CFABSOLUTETIME_INVALID;
         _startClock = currentMonotonicClockNsecIfUnset(startAbsTime);
         _name = name;
         _parentId = parentId;
@@ -366,6 +368,7 @@ static CFAbsoluteTime currentTimeIfUnset(CFAbsoluteTime time) {
         copy.kind = self.kind;
         copy.isMutable = self.isMutable;
         copy.hasBeenProcessed = self.hasBeenProcessed;
+        copy.isAppSessionSpan = self.isAppSessionSpan;
         copy.wasStartOrEndTimeProvided = self.wasStartOrEndTimeProvided;
         copy.state = self.state;
         copy.startClock = self.startClock;